IRISSYNCS-9 sync 与 node 断开连接后的重试机制 1:支持tcp连接心跳检测 2:支持动态添加tcp连接到连接池 3:支持节点发现，实现高可用

+++ util/helper/pool_client_test.go
@@ -34,7 +34,7 @@ func TestClient_Release(t *testing.T) {
 	type fields struct {
 		Client client.Client
 		used   bool
-		id     int64
+		id     int
 	}
 	tests := []struct {
 		name   string
